Add content_length formatting benches
