--- crates/fluke/benches/format_status_code.rs
@@ -1,54 +1,0 @@
-use codspeed_criterion_compat::{black_box, criterion_group, criterion_main, Criterion};
-use http::StatusCode;
-
-pub fn format_status_code(c: &mut Criterion) {
-    let status_codes = (100..=999)
-        .map(|code| StatusCode::from_u16(code).unwrap())
-        .collect::<Vec<StatusCode>>();
-
-    let mut c = c.benchmark_group("format_status_code");
-
-    c.bench_function("std::fmt", |b| {
-        b.iter_batched(
-            || status_codes.clone(),
-            |codes| {
-                for code in &codes {
-                    black_box(code.as_u16().to_string());
-                }
-            },
-            codspeed_criterion_compat::BatchSize::SmallInput,
-        )
-    });
-
-    c.bench_function("itoa", |b| {
-        b.iter_batched(
-            || status_codes.clone(),
-            |codes| {
-                for code in &codes {
-                    use itoa::Buffer;
-                    let mut buffer = Buffer::new();
-                    black_box(buffer.format(code.as_u16()).to_string());
-                }
-            },
-            codspeed_criterion_compat::BatchSize::SmallInput,
-        )
-    });
-
-    c.bench_function("lookup_table", |b| {
-        b.iter_batched(
-            || status_codes.clone(),
-            |codes| {
-                for code in &codes {
-                    let code = black_box(*code);
-                    black_box(fluke::h1::encode::encode_status_code(code));
-                }
-            },
-            codspeed_criterion_compat::BatchSize::SmallInput,
-        )
-    });
-
-    c.finish()
-}
-
-criterion_group!(benches, format_status_code);
-criterion_main!(benches);
